enhancement: make decimal optional on minting
