+++ packages/shared/components/inputs/index.js
@@ -5,7 +5,7 @@ export { default as AssetAmountInput } from './AssetAmountInput.svelte'
 export { default as AssetDropdown } from './AssetDropdown.svelte'
 export { default as Checkbox } from './Checkbox.svelte'
 export { default as ClosableInputContainer } from './ClosableInputContainer.svelte'
-export { default as ClosableTextInput } from './ClosableTextInput.svelte'
+export { default as ClosableInput } from './ClosableInput.svelte'
 export { default as DateTimePicker } from './DateTimePicker.svelte'
 export { default as Dropdown } from './Dropdown.svelte'
 export { default as Dropdown2 } from './Dropdown2.svelte'
